Update PFC_TrialEvent_MLB_SM.m
Adjusted run conditions to allow for creation of object without running the analysis (necessary for updating from default parameters or in this case identifying a unit exclusion vector for units).

Also realized I could cut out the individual lines in the creation and just replace it with a RunAnalysis call because it's just doing the same thing.